Add WallflyCalendar component, compose it in Calendar
